pt-119133379: token grant can ask for a token

+++ src/main/java/org/rootservices/authorization/oauth2/grant/redirect/code/token/MakeToken.java
@@ -8,6 +8,6 @@ import org.rootservices.authorization.persistence.entity.Token;
  */
 public interface MakeToken {
     Token run(String plainTextToken);
-    Integer getSecondsToExpiration();
+    Long getSecondsToExpiration();
     TokenType getTokenType();
 }
